Safari-609.1.11.1/LayoutTests/webgpu/whlsl/double-not.html

<!DOCTYPE html>
<html>
<meta charset=utf-8>
<meta name="timeout" content="long">
<title>Copying.</title>
<script src="js/test-harness.js"></script>
<script src="../js/webgpu-functions.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
const whlslTests = {};

whlslTests.doubleNot = async () =>
{
    let program = `
        bool foo(bool x)
        {
            return !!x;
        }
    `;
    assert_equals(await callBoolFunction(program, "foo", [makeBool(true)]), true);
    assert_equals(await callBoolFunction(program, "foo", [makeBool(false)]), false);
}

runTests(whlslTests);
</script>
</html>
